backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.0.1
open-clip-torch==2.24.0
faiss-cpu==1.7.4
numpy==1.24.3
pillow==10.0.1
python-multipart==0.0.6
pydantic==2.4.2
httpx==0.25.0
